Move components to separate files
